% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_apis.R
\name{agsi_gas_data}
\alias{agsi_gas_data}
\title{Fetch Gas Data from AGSI API}
\usage{
agsi_gas_data(country, from_date, to_date, api_key = Sys.getenv("AGSI_KEY"))
}
\arguments{
\item{country}{Character. Country code for which the data is to be fetched (e.g., "IT").}

\item{from_date}{Character. Start date for the data retrieval in "YYYY-MM-DD" format.}

\item{to_date}{Character. End date for the data retrieval in "YYYY-MM-DD" format.}

\item{api_key}{Character. API key for authenticating with the AGSI API.}
}
\value{
A data.table containing gas data for the specified country and date range.
If the API request fails, an empty data.table is returned.
}
\description{
This function retrieves gas data for a specified country and date range using the AGSI API.
The results are returned as a data.table with detailed gas storage information.
}
\details{
The function fetches data from the AGSI API, iterating over multiple pages if necessary.
It consolidates the results into a single data.table for further analysis.
}
\examples{
\dontrun{
api_key <- 'your_api_key_here'
country <- "IT"
from_date <- "2023-10-01"
to_date <- "2023-12-01"
data <- fetch_country_data(api_key, country, from_date, to_date)
}

}
