% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_apis.R
\name{alba_download_data_gas}
\alias{alba_download_data_gas}
\title{Download and process gas price data from GEEO Energy}
\usage{
alba_download_data_gas(
  from_date,
  to_date,
  type,
  output_dir,
  username,
  password
)
}
\arguments{
\item{from_date}{A Date object or a character string in "YYYY-MM-DD" format specifying the start date for filtering the data.}

\item{to_date}{A Date object or a character string in "YYYY-MM-DD" format specifying the end date for filtering the data.}

\item{output_dir}{A character string specifying the output directory where the file will be saved.}

\item{username}{A character string specifying the username for login.}

\item{password}{A character string specifying the password for login.}
}
\value{
A \code{data.table} containing the processed gas price data.
}
\description{
This function logs into the GEEO Energy website, downloads the gas price data,
processes the data using the \code{process_xlsx} function, and then deletes the
downloaded file after processing.
}
\details{
This function logs into the GEEO Energy platform, retrieves the gas price data as an Excel file,
processes it to extract relevant data within the specified date range, and then deletes the file
after processing.
}
