% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_req}
\alias{api_req}
\title{Make an API Request to ENTSO-E Web API}
\usage{
api_req(
  api_scheme = "https://",
  api_domain = "web-api.tp.entsoe.eu/",
  api_name = "api?",
  query_string = NULL,
  api_key = NULL
)
}
\arguments{
\item{api_scheme}{A string representing the API scheme, default is \code{"https://"}. This is the protocol used for the API request.}

\item{api_domain}{A string representing the domain name of the API, default is \code{"web-api.tp.entsoe.eu/"}.}

\item{api_name}{A string representing the API name or endpoint, default is \code{"api?"}.}

\item{query_string}{A string representing the query parameters to be appended to the API URL. This parameter is required.}

\item{api_key}{A string representing the security token used for authentication. This parameter is required.}
}
\value{
A list containing the response data:
\describe{
\item{en_cont_list}{A list of XML content if the response is a zip file. The list contains the decompressed XML files.}
\item{en_cont}{XML content if the response is an XML document.}
}
}
\description{
This function sends a request to the ENTSO-E (European Network of Transmission System Operators for Electricity) Web API, retrieves data based on the given query, and processes the response. It handles both XML and zip file responses.
}
\details{
The function first validates the presence of the required arguments: \code{query_string} and \code{api_key}. If the API request is successful, it handles both XML responses and zip file responses:
\itemize{
\item If the response is a zip file (\code{application/zip}), it downloads and extracts the XML content.
\item If the response is XML (\code{text/xml} or \code{application/xml}), it returns the XML content directly.
}

If the request fails with a \verb{200 OK} response but contains an error message or exceeds the data limit, the function checks for the need to offset the query and recursively retries with a new query offset.
}
\examples{
\dontrun{
# Example usage of the api_req function
query_string <- "documentType=A63&businessType=A85"
api_key <- "your_api_key"
result <- api_req(query_string = query_string, api_key = api_key)
print(result)
}

}
