% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motherduck.R
\name{check_database_md}
\alias{check_database_md}
\title{Check if a MotherDuck Database Exists}
\usage{
check_database_md(connection, database_name)
}
\arguments{
\item{connection}{A valid DuckDB connection object, typically returned by
\code{DBI::dbConnect()}. This connection must be established and remain active.}

\item{database_name}{A character string representing the alias of the
database you want to check.}
}
\value{
A logical value:
\itemize{
\item \code{TRUE} if the specified \code{database_name} exists in the MotherDuck databases.
\item \code{FALSE} if the database does not exist or the connection is invalid.
}
}
\description{
This function checks whether a specified database exists in the connected
MotherDuck environment using an active DuckDB connection.
}
\details{
The function queries the list of all available databases using the
\code{MD_ALL_DATABASES()} function in MotherDuck, then filters the results for
active, attached databases of type 'motherduck'. If the \code{database_name} exists
in the list, the function returns \code{TRUE} and prints a success message. Otherwise,
it returns \code{FALSE} and notifies the user that the database does not exist.

If the connection object is invalid or closed, the function will return
\code{FALSE} and print an error message.
}
\examples{
\dontrun{
  # Assuming you have an active connection named 'conn'
  check_database_md(conn, "my_database")
}

}
