% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motherduck.R
\name{check_table_md}
\alias{check_table_md}
\title{Check if a Table Exists in a MotherDuck Database}
\usage{
check_table_md(connection, table_name, verbose = TRUE)
}
\arguments{
\item{connection}{A valid DuckDB connection object, typically returned by
\code{DBI::dbConnect()}. This connection must be established and remain active.}

\item{table_name}{A character string representing the name of the table you want to check.}

\item{verbose}{A bolean to print messages}
}
\value{
A logical value:
\itemize{
\item \code{TRUE} if the specified \code{table_name} exists in the connected database.
\item \code{FALSE} if the table does not exist or the connection is invalid.
}
}
\description{
This function checks whether a specified table exists in the connected
MotherDuck environment using an active DuckDB connection.
}
\details{
The function retrieves all tables in the currently connected database using
\code{DBI::dbListTables()}. If the specified \code{table_name} exists in the list of
tables, the function returns \code{TRUE} and prints a success message. Otherwise,
it returns \code{FALSE} and notifies the user that the table does not exist.

If the connection object is invalid or closed, the function will return
\code{FALSE} and print an error message.
}
\examples{
\dontrun{
  # Assuming you have an active connection named 'conn'
  check_table_md(conn, "my_table")
}

}
