% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motherduck.R
\name{connect_md}
\alias{connect_md}
\title{Connect to a MotherDuck Database}
\usage{
connect_md(store_conn = TRUE)
}
\arguments{
\item{store_conn}{Logical, default is \code{TRUE}. If \code{TRUE}, the function stores the
established DuckDB connection object (\code{conn}) in the global environment. If
\code{FALSE}, it does not store the connection.}
}
\value{
The function returns no value but will optionally store the \code{conn} object
in the global environment if \code{store_conn = TRUE}. On failure, error messages
are printed, and the function will stop execution if the error is not related
to authentication.
}
\description{
This function establishes a connection to a local DuckDB database and connects
to the MotherDuck API using the authentication token stored in the environment
variable \code{MOTHERDUCK}. It optionally stores the connection in the global
environment for future use.
}
\details{
The function also handles authentication errors and informs the user if the
connection was successful or not.

The function first establishes a local DuckDB connection and installs and loads
the \code{motherduck} extension. It then sets the MotherDuck token using the token
stored in the \code{MOTHERDUCK} environment variable. If the connection is successful,
it confirms the connection status and optionally stores the connection.
\itemize{
\item If the MotherDuck token is missing or invalid, the function returns an error message.
\item If the connection is successful, a message indicating success is displayed.
}
}
\examples{
\dontrun{
  # Connect and store the connection globally
  connect_md(TRUE)

  # Connect without storing the connection globally
  connect_md(FALSE)
}

}
