% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aux.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country code ISO 3166}
\format{
A data frame with X rows and Y variables:
\describe{
\item{country}{Country description}
\item{code_2}{Alpha-2 digit code ISO 3166}
\item{code_3}{Alpha-3 digit code ISO 3166}
\item{code_num}{Numeric code ISO 3166}
}
}
\source{
\url{https://www.iban.com/country-codes}
}
\usage{
country_codes
}
\description{
A dataset containing the alpha-2 and alpha-3 digits code for countries to be used.
}
\keyword{datasets}
