% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motherduck.R
\name{create_empty_table_md}
\alias{create_empty_table_md}
\title{Create an Empty Table in MotherDuck Database Dynamically}
\usage{
create_empty_table_md(conn, database_name, table_name, data)
}
\arguments{
\item{conn}{A valid DuckDB connection object.}

\item{database_name}{The name of the MotherDuck database where the table will be created.}

\item{table_name}{A character string representing the name of the new table.}

\item{data}{A data frame or data table (\code{dt_all[[i]]}) from which column names and types are derived.}
}
\value{
TRUE if the table is successfully created, FALSE otherwise.
}
\description{
This function creates a new table in a specified MotherDuck database
based on the column names and types of the input data.
}
\examples{
\dontrun{
  create_empty_table_md(conn, "economics_db", "new_table", mtcars)
}
}
