% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_download_extracted_file}
\alias{datascope_download_extracted_file}
\title{Download Extracted File from Refinitiv Datascope}
\usage{
datascope_download_extracted_file(
  extracted_file_id,
  auth_token,
  filename = "raw.csv"
)
}
\arguments{
\item{extracted_file_id}{Character. The unique identifier of the extracted file.}

\item{auth_token}{Character. The authentication token for accessing the API.}

\item{save_path}{Character. The local file path where the downloaded file will be saved.}
}
\value{
Character. The path where the file has been saved.
}
\description{
This function downloads an extracted file from the Refinitiv Datascope API using the provided
\code{extracted_file_id} and \code{auth_token}. The file is saved to the specified \code{save_path}.
}
\examples{
\dontrun{
download_extracted_file("VjF8fDMzNjE5Mjk", "<your_auth_token>", "CM_STRIKE_PRICE_20250329.csv")
}

}
