% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_instrument_items}
\alias{datascope_instrument_items}
\title{Retrieve Items from a Specific Instrument List in Refinitiv DataScope Select API}
\usage{
datascope_instrument_items(instrument_list_id, auth_token)
}
\arguments{
\item{instrument_list_id}{A character string representing the ID of the
instrument list from which to retrieve items.}

\item{auth_token}{A character string representing the authentication token
obtained from the Refinitiv DataScope Select API.}
}
\value{
A data table containing the items from the specified instrument list.
}
\description{
This function sends a \code{GET} request to the Refinitiv DataScope Select API
to retrieve items from a specific instrument list based on the provided
instrument list ID and authentication token. It processes the response
and returns the result as a data table.
}
\examples{
\dontrun{
  instrument_list_id <- "'0x08532ebecc787a72'"
  auth_token <- "<your_auth_token>"
  instrument_items <- get_instrument_items(instrument_list_id, auth_token)
  print(instrument_items)
}
}
