% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_report_extraction_files}
\alias{datascope_report_extraction_files}
\title{Retrieve Files for a Specific Report Extraction from Refinitiv DataScope Select API}
\usage{
datascope_report_extraction_files(
  report_extraction_id,
  auth_token,
  file_type = ".csv"
)
}
\arguments{
\item{report_extraction_id}{A character string representing the ID of the
report extraction from which to retrieve files.}

\item{auth_token}{A character string representing the authentication token
obtained from the Refinitiv DataScope Select API.}
}
\value{
A data table containing the files for the specified report extraction.
}
\description{
This function sends a \code{GET} request to the Refinitiv DataScope Select API
to retrieve the files associated with a specific report extraction,
based on the provided report extraction ID and authentication token.
It processes the response and returns the result as a data table.
}
\examples{
\dontrun{
  report_extraction_id <- "'6651000'"
  auth_token <- "<your_auth_token>"
  report_files <- get_report_extraction_files(report_extraction_id, auth_token)
  print(report_files)
}
}
