% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_report_lastextraction}
\alias{datascope_report_lastextraction}
\title{Retrieve Last Extraction for a Specific Schedule from Refinitiv DataScope Select API}
\usage{
datascope_report_lastextraction(schedule_id, auth_token)
}
\arguments{
\item{schedule_id}{A character string representing the ID of the schedule.}

\item{auth_token}{A character string representing the authentication token
obtained from the Refinitiv DataScope Select API.}
}
\value{
A data table containing the last extraction details for the specified schedule.
}
\description{
This function sends a \code{GET} request to the Refinitiv DataScope Select API
to retrieve the last extraction for a specific schedule, based on the provided
schedule ID and authentication token. It processes the response and returns
the result as a data table.
}
\examples{
\dontrun{
  schedule_id <- "'0x05807049631b1f86'"
  auth_token <- "<your_auth_token>"
  last_extraction <- get_last_extraction(schedule_id, auth_token)
  print(last_extraction)
}
}
