% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_report_template_schedules}
\alias{datascope_report_template_schedules}
\title{Retrieve Schedules for a Specific Report Template from Refinitiv DataScope Select API}
\usage{
datascope_report_template_schedules(report_template_id, auth_token)
}
\arguments{
\item{report_template_id}{A character string representing the ID of the
report template from which to retrieve schedules.}

\item{auth_token}{A character string representing the authentication token
obtained from the Refinitiv DataScope Select API.}
}
\value{
A data table containing the schedules for the specified report template.
}
\description{
This function sends a \code{GET} request to the Refinitiv DataScope Select API
to retrieve the schedules associated with a specific report template,
based on the provided report template ID and authentication token.
It processes the response and returns the result as a data table.
}
\examples{
\dontrun{
  report_template_id <- "'0x12345abcde'"
  auth_token <- "<your_auth_token>"
  report_schedules <- get_report_template_schedules(report_template_id, auth_token)
  print(report_schedules)
}
}
