% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_report_templates}
\alias{datascope_report_templates}
\title{Retrieve All Report Templates from Refinitiv DataScope Select API}
\usage{
datascope_report_templates(auth_token)
}
\arguments{
\item{auth_token}{A character string representing the authentication token
obtained from the Refinitiv DataScope Select API.}
}
\value{
A data table containing the available report templates.
}
\description{
This function sends a \code{GET} request to the Refinitiv DataScope Select API
to retrieve all available report templates. It processes the response
and returns the result as a data table.
}
\examples{
\dontrun{
  auth_token <- "<your_auth_token>"
  report_templates <- get_report_templates(auth_token)
  print(report_templates)
}
}
