% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascope.R
\name{datascope_token}
\alias{datascope_token}
\title{Authenticate with Refinitiv DataScope Select API}
\usage{
datascope_token(username, password)
}
\arguments{
\item{username}{A character string representing the Refinitiv username.}

\item{password}{A character string representing the Refinitiv password.}
}
\value{
A character string representing the authentication token.
}
\description{
This function sends a \code{POST} request to the Refinitiv DataScope Select API
to obtain an authentication token using the provided username and password.
}
\examples{
\dontrun{
  token <- authenticate_refinitiv("9028810", "Ref-e2021")
  print(token)
}
}
