% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_countries.R
\docType{data}
\name{entsoe_countries}
\alias{entsoe_countries}
\title{List of Mapped Countries for DAM Prices}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 43 rows and 2 columns.
}
\usage{
entsoe_countries
}
\description{
This data table contains a mapping of countries to their respective EIC codes for retrieving DAM prices from the ENTSOE Transparency Platform API.
}
\examples{
library(entsoeapi)
entsoe_countries

}
\keyword{datasets}
