% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoeftp.R
\name{entsoe_create_url_folders}
\alias{entsoe_create_url_folders}
\title{Create URL Folder Request to ENTSO-E SFTP Server}
\usage{
entsoe_create_url_folders(
  user = Sys.getenv("ENTSOE_USER"),
  psw = Sys.getenv("ENTSOE_PASSWORD")
)
}
\arguments{
\item{user}{A character string with the username for ENTSO-E SFTP server authentication.}

\item{psw}{A character string with the password for ENTSO-E SFTP server authentication.}
}
\value{
A response object containing the contents of the specified SFTP directory,
as retrieved from the ENTSO-E transparency platform.
}
\description{
This function creates and sends a request to the ENTSO-E SFTP server's transparency platform,
retrieving the list of available folders or files. It authenticates using provided credentials
(\code{user} and \code{psw}) for secure FTP access.
}
\details{
This function uses \code{curl} to handle the HTTP request for secure FTP access.
Ensure that the \code{user} and \code{psw} parameters are provided with valid ENTSO-E credentials.
}
\examples{
\dontrun{
# Example usage:
entsoe_create_url_folders(user = "your_username", psw = "your_password")
}

}
