% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_sftp.R
\name{entsoe_download_file}
\alias{entsoe_download_file}
\title{Download a File from ENTSO-E SFTP Server}
\usage{
entsoe_download_file(
  basis_name,
  year,
  month,
  user = Sys.getenv("ENTSOE_USER"),
  password = Sys.getenv("ENTSOE_PASSWORD"),
  output_file
)
}
\arguments{
\item{basis_name}{A character string specifying the high-level folder name of the file.}

\item{year}{An integer specifying the year of the file to retrieve.}

\item{month}{An integer specifying the month of the file to retrieve (e.g., 1 for January).}

\item{user}{A character string for the username for SFTP authentication.}

\item{password}{A character string for the password for SFTP authentication.}

\item{output_file}{A character string specifying the local file path where the downloaded file will be saved.}
}
\value{
Logical value indicating whether the file was downloaded successfully.
}
\description{
This function downloads a specified file from the ENTSO-E SFTP server using the provided credentials.
}
\examples{
\dontrun{
# Download a file for specified parameters
download_successful <- download_entsoe_file(basis_name = "example_folder", year = 2023, month = 1,
                                              user = Sys.getenv("ENTSOE_USER"),
                                              password = Sys.getenv("ENTSOE_PASSWORD"),
                                              output_file = "local_file_path.csv")
}

}
