% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_api.R
\name{entsoe_generation_by_type}
\alias{entsoe_generation_by_type}
\title{Retrieve Actual Generation by Production Type for a Specific Country}
\usage{
entsoe_generation_by_type(
  country,
  from_data,
  to_data,
  api_key = Sys.getenv("ENTSOE_KEY")
)
}
\arguments{
\item{country}{A character string specifying the country for which to retrieve generation data. The country should match an entry in the \code{entsoe_countries} table.}

\item{from_data}{A Date object specifying the start date for the data range. Default is 365 days before the current date.}

\item{to_data}{A Date object specifying the end date for the data range. Default is the current date.}

\item{api_key}{A character string containing the ENTSOE API key. It can be passed directly or retrieved from the environment variable \code{ENTSOE_KEY}.}
}
\value{
A \code{data.table} containing the formatted generation data for the specified country and time range. The data table includes:
\itemize{
\item \code{DATE}: The date in \code{YYYY-MM-DD} format.
\item \code{HOUR}: The hour of the day (position in the data).
\item \code{PRODUCTION_TYPE}: The type of production (e.g., solar, wind, nuclear).
\item \code{VALUE}: The generation value in MW.
\item \code{UNIT}: The unit of the generation value (always 'MW').
}
}
\description{
This function retrieves actual generation by production type data for a specific country from the ENTSOE Transparency Platform API.
The function fetches generation data for a given time range and country, processes the data, and formats it into a \code{data.table}.
}
\examples{
# Example usage to retrieve generation data for Italy (North)
country <- "Italy (North)"
from_data <- Sys.Date() - 365
to_data <- Sys.Date()
generation_data <- entsoe_generation_by_type(country, from_data, to_data, api_key = Sys.getenv('ENTSOE_KEY'))
print(generation_data)

}
