% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoeftp.R
\name{entsoe_list_files}
\alias{entsoe_list_files}
\title{List Files within a Specified Folder on ENTSO-E SFTP Server}
\usage{
entsoe_list_files(basis_name)
}
\arguments{
\item{basis_name}{A character string specifying the name of the folder from which to retrieve the file listing.}
}
\value{
A data frame containing information about files within the specified folder on the ENTSO-E SFTP server.
The data frame includes file names and any other available metadata.
}
\description{
This function retrieves and lists files from a specified folder (\code{basis_name}) on the ENTSO-E SFTP server's transparency platform.
It uses the \code{entsoe_create_url_files} function to fetch the file listing and processes the raw content with \code{parse_ftp_files}.
}
\details{
The function sends a request to the ENTSO-E SFTP server, retrieves the list of files as raw text,
and converts it into a structured format using \code{parse_ftp_files}. Ensure that \code{basis_name} corresponds to an
existing folder on the server.
}
\examples{
\dontrun{
# List files in a folder named "example_folder":
head(entsoe_list_files(basis_name = "OutagesPU"))
}

}
