% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_sftp.R
\name{entsoe_list_folders}
\alias{entsoe_list_folders}
\title{List Top-Level Folder Links from ENTSO-E SFTP Server}
\usage{
entsoe_list_folders()
}
\value{
A data frame containing the names and metadata of top-level folders on the ENTSO-E SFTP server.
}
\description{
This function retrieves and lists the top-level folder links from the ENTSO-E SFTP server's transparency platform.
It uses the \code{entsoe_create_url_folders} function to fetch the raw folder list and processes it using \code{parse_ftp_links}.
}
\details{
The function makes a request to the ENTSO-E SFTP server, retrieves folder content as raw text,
and converts it into a structured format using \code{parse_ftp_links}. The resulting data frame lists folder names
and relevant details.
}
\examples{
\dontrun{
# List the folders:
head(entsoe_list_folders())
}

}
