% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{find_missing_dates}
\alias{find_missing_dates}
\title{Find Missing Dates in a Vector}
\usage{
find_missing_dates(dates, full_week = TRUE)
}
\arguments{
\item{dates}{A character vector of dates in "YYYY-MM-DD" format.}

\item{full_week}{Logical. If \code{TRUE}, includes all missing dates (default). If \code{FALSE}, only includes missing weekdays (Monday to Friday).}
}
\value{
A character vector of missing dates in "YYYY-MM-DD" format.
}
\description{
This function identifies the missing dates in a sequence of dates. It can optionally exclude weekends (Saturday and Sunday) from the results.
}
\examples{
# Example input vector
ret_files <- c(
  "2023-12-31", "2024-12-01", "2024-12-02", "2024-12-03", "2024-12-04",
  "2024-12-14", "2024-12-20", "2024-12-21", "2024-12-22", "2024-12-23",
  "2024-12-24", "2024-12-25", "2024-12-26", "2024-12-27", "2024-12-28",
  "2024-12-29"
)

# Find all missing dates (full week)
find_missing_dates(ret_files, full_week = TRUE)

# Find only missing weekdays (Monday to Friday)
find_missing_dates(ret_files, full_week = FALSE)

}
