% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_api.R
\name{fun_fields}
\alias{fun_fields}
\title{Generate MongoDB Projection String}
\usage{
fun_fields(v_field)
}
\arguments{
\item{v_field}{A character vector of field names to include in the projection.}
}
\value{
A JSON string representing the MongoDB projection.
}
\description{
This function generates a MongoDB projection string from a vector of field names.
}
