% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_api.R
\name{fun_query_date}
\alias{fun_query_date}
\title{Generate MongoDB Date Range Query}
\usage{
fun_query_date(field_name, start_date, end_date)
}
\arguments{
\item{field_name}{The name of the date field to query.}

\item{start_date}{The start date of the range.}

\item{end_date}{The end date of the range.}
}
\value{
A JSON string representing the MongoDB date range query.
}
\description{
This function generates a MongoDB query string for a date range.
}
