% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_api.R
\name{fun_query_fl}
\alias{fun_query_fl}
\title{Generate MongoDB Query for List of Values}
\usage{
fun_query_fl(field_name, v_values)
}
\arguments{
\item{field_name}{The name of the field to query.}

\item{v_values}{A vector of values to include in the query.}
}
\value{
A JSON string representing the MongoDB query for the list of values.
}
\description{
This function generates a MongoDB query string for a list of values.
}
