% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest_date}
\alias{get_latest_date}
\title{Get the Latest Date from a MongoDB Collection}
\usage{
get_latest_date(collection)
}
\arguments{
\item{collection}{An object representing the MongoDB collection connection.}
}
\value{
A list containing the latest date (\code{DATE}) and the associated value (\code{VALUE}), or \code{NULL} if the collection is empty.
}
\description{
This function retrieves the latest record by date from a MongoDB collection.
}
