% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_download.R
\name{gme_igi_get_files}
\alias{gme_igi_get_files}
\title{Retrieve Files from GME FTP Server for DAM market}
\usage{
gme_igi_get_files(
  data_type,
  output_dir = "data",
  username = "PIASARACENO",
  password = "18N15C9R",
  verbose = FALSE,
  storico = FALSE,
  n = 1
)
}
\arguments{
\item{data_type}{A string specifying the data type directory to query.
Default is "MGP_Prezzi". Can be customized for other directories.}

\item{output_dir}{A string specifying the directory where output files will be saved.
Default is "data". If the directory does not exist, it will be created.}

\item{username}{FTP server username. Default is "PIASARACENO".}

\item{password}{FTP server password. Default is "18N15C9R".}

\item{verbose}{Logical. If TRUE, prints the list of available files. Default is FALSE.}
}
\value{
A character vector containing the filenames that match the specified pattern,
or NULL if an error occurs.
}
\description{
This function connects to the GME FTP server to list XML files available in a specified directory.
It can filter files based on the provided \code{data_type}.
}
\examples{
\dontrun{
files <- gme_mgp_get_files(data_type = "MGP_Prezzi", verbose = TRUE)
}
}
