% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_download.R
\name{gme_mi_get_files}
\alias{gme_mi_get_files}
\title{Retrieve List of Available GME MI Market XML Files from FTP}
\usage{
gme_mi_get_files(
  data_type,
  output_dir = "data",
  username = "PIASARACENO",
  password = "18N15C9R",
  verbose = FALSE
)
}
\arguments{
\item{data_type}{Character string specifying the type of data to retrieve. Must be one of:
\code{"MI-A1_Prezzi"}, \code{"MI-A2_Prezzi"}, \code{"MI-A3_Prezzi"},
\code{"MI-A1_Quantita"}, \code{"MI-A2_Quantita"}, \code{"MI-A3_Quantita"}.}

\item{output_dir}{Directory path to store the files (not used in this version, but directory is ensured to exist).
Defaults to \code{"data"}.}

\item{username}{FTP username. Default is \code{"PIASARACENO"}.}

\item{password}{FTP password. Default is \code{"18N15C9R"}.}

\item{verbose}{Logical; if \code{TRUE}, prints the list of matched files. Default is \code{FALSE}.}
}
\value{
A character vector of filenames matching the \code{data_type} on the FTP server.
Returns \code{NULL} in case of an error.
}
\description{
Connects to the GME FTP server and lists available XML files for MI-A1, MI-A2, or MI-A3
market sessions (prices or quantities). Filters files by the appropriate naming pattern.
}
\examples{
\dontrun{
files <- gme_mi_get_files(data_type = "MI-A1_Prezzi", verbose = TRUE)
}

}
