% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_flux_entsoe}
\alias{install_flux_entsoe}
\title{Install the \code{flux_lseg} Python Package in a Virtual Environment}
\usage{
install_flux_entsoe(env = NULL)
}
\arguments{
\item{env}{Optional. The name or path of an existing Python virtual environment.
If \code{NULL}, a new virtual environment named \code{"fluxenv"} is created using the default Python.}
}
\value{
No return value. The function performs the installation as a side effect and prints a success message.
}
\description{
Installs the bundled \code{flux_lseg} Python package (provided as a \code{.tar.gz} file
under \verb{inst/python/}) into a Python virtual environment using the \code{reticulate} package.
}
\details{
included in the R package under \verb{inst/python/}. It uses \code{reticulate::virtualenv_install()} to
install the package via \code{pip}.

Python 3 must be available. You can specify the Python path by setting the \code{RETICULATE_PYTHON}
environment variable or using \code{reticulate::use_python()} prior to calling this function.
}
