% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gme_ftp.R
\name{mgp_download_file}
\alias{mgp_download_file}
\title{Download and Process MGP Data File}
\usage{
mgp_download_file(
  filename,
  username,
  password,
  output_dir,
  data_type = "MGP_Prezzi",
  raw = FALSE
)
}
\arguments{
\item{filename}{A character string representing the name of the file to be downloaded.}

\item{username}{A character string representing the FTP username for authentication.}

\item{password}{A character string representing the FTP password for authentication.}

\item{output_dir}{A character string representing the directory where the downloaded file will be saved.}
}
\value{
A data frame (or \code{NULL} in case of an error). The data frame contains the processed data
obtained from the downloaded XML file, or \code{NULL} if an error occurred during download or processing.
}
\description{
This function downloads a file from an FTP server using provided credentials, processes the
XML data from the downloaded file, and returns the processed data. After processing, the
downloaded XML file is deleted.
}
\details{
This function uses the \code{curl} package to download the file from an FTP server with the provided
credentials. After downloading, it passes the downloaded file to the \code{gme_dam_xml_to_data} function
for processing. Finally, the downloaded XML file is deleted from the local system to clean up.
}
\examples{
# Example usage:
result <- mgp_download_file("example.xml", "your_username", "your_password", "/path/to/output_dir")
print(result)

}
