% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aux.R
\docType{data}
\name{mongo_aggregationverbs}
\alias{mongo_aggregationverbs}
\title{Mongo Aggregation Verbs}
\format{
A data.table with two columns:
\describe{
\item{v_aggregation_short}{Short names for aggregation functions.}
\item{v_aggregation_long}{Corresponding long names for aggregation functions.}
}
}
\usage{
mongo_aggregationverbs
}
\description{
This data table maps short names of aggregation functions to their corresponding long names.
}
\examples{
mongo_aggregationverbs
#    v_aggregation_short v_aggregation_long
# 1:                 sum                 Sum
# 2:                mean             Average
# 3:                 max            Maximum
# 4:                 min            Minimum
# 5:              length               Count

}
\keyword{datasets}
