% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_apis.R
\name{process_alba_xlsx}
\alias{process_alba_xlsx}
\title{Process Excel file and extract relevant data}
\usage{
process_alba_xlsx(filename, from_date, to_date, type)
}
\arguments{
\item{filename}{A character string specifying the path to the Excel file to be processed.}

\item{from_date}{A Date object or a character string in "YYYY-MM-DD" format specifying the start date for filtering the data.}

\item{to_date}{A Date object or a character string in "YYYY-MM-DD" format specifying the end date for filtering the data.}

\item{type}{A character string specifying the type of commodity. It can be either "gas" or "electricity".}
}
\value{
A \code{data.table} containing the processed data, with columns:
\itemize{
\item \code{DATE}: The date of the observation.
\item \code{TYPE}: The commodity type (e.g., "PSV", "COAL").
\item \code{VARIABLE}: The name of the variable (e.g., "PSV MW").
\item \code{VALUE}: The value of the variable for the given date.
}
}
\description{
This function processes an Excel file containing multiple sheets, extracts
relevant data based on the type of commodity (gas or electricity), filters
the data within the specified date range, and combines it into a single
\code{data.table}. The data is reshaped, cleaned, and additional columns are added
for further analysis.
}
\details{
This function processes the following sheets for "gas" type: "PSV MW", "GR04 MW", "TTF MW",
"GR07 MW", "Baum MW", "NCG MW", "PEG MW". For "electricity" type, the following sheets
are processed: "Elettricita Italia", "Elettricita Francia", "Elettricita Germania", "COAL",
and "CO2". Data is filtered by the \code{from_date} and \code{to_date} parameters, reshaped from wide
to long format, and combined into a single \code{data.table}.
}
\examples{
# Example usage:
filename <- "path/to/your/file.xlsx"
from_date <- Sys.Date() - 7
to_date <- Sys.Date()
type <- "electricity"
processed_data <- process_xlsx(filename, from_date, to_date, type)
}
