% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_api.R
\name{query_mongo}
\alias{query_mongo}
\title{Query MongoDB Data API}
\usage{
query_mongo(
  collection_name,
  filter_list = NULL,
  sort_list = list(DATE = -1),
  cluster = "Cluster0",
  db_name = "ECONOMICS",
  output = "SIMPLE",
  print_output = FALSE
)
}
\arguments{
\item{collection_name}{A character string specifying the name of the MongoDB collection. This parameter is required.}

\item{filter_list}{A named list of filters to apply to the query. Defaults to \code{NULL} to retrieve all documents.}

\item{sort_list}{A named list of sorting options. Defaults to \code{list("DATE" = -1)} to sort by "DATE" in descending order.}

\item{cluster}{A character string specifying the cluster name. Defaults to 'Cluster0'.}

\item{db_name}{A character string specifying the database name. Defaults to 'ECONOMICS'.}

\item{output}{A character string specifying the type of output: 'SIMPLE' for a data table or 'RAW' for the raw response. Defaults to 'SIMPLE'.}

\item{print_output}{A logical value indicating whether to print the output. Defaults to TRUE.}
}
\value{
A data table containing the query results if \code{output} is 'SIMPLE', or the raw response if \code{output} is 'RAW'.
}
\description{
This function queries a MongoDB collection using the MongoDB Data API.
}
