% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{to_yymm}
\alias{to_yymm}
\title{Convert Dates to First Day of the Month}
\usage{
to_yymm(dates)
}
\arguments{
\item{dates}{A vector of dates in \code{"YYYY-MM-DD"} format. Can be \code{character}, \code{Date}, or \code{IDate}.}
}
\value{
A vector of dates converted to the first day of the respective month (\code{IDate} format).
}
\description{
This function transforms a given date (or vector of dates) into the first day of the same month, formatted as \code{YYYY-MM-01}.
}
