% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe_sftp.R
\name{entsoe_actual_generation}
\alias{entsoe_actual_generation}
\title{Retrieve and Process ENTSO-E Actual Generation Data}
\usage{
entsoe_actual_generation(
  year_data,
  month_data,
  raw = FALSE,
  user = Sys.getenv("ENTSOE_USER"),
  password = Sys.getenv("ENTSOE_PASSWORD"),
  output_folder = tempdir()
)
}
\arguments{
\item{year_data}{Numeric. The year for which to download the ENTSO-E data (e.g., 2024).}

\item{month_data}{Numeric. The month for which to download the ENTSO-E data (e.g., 10 for October).}

\item{raw}{Logical. If \code{TRUE}, keeps the downloaded CSV file after processing. If \code{FALSE}, the CSV file is deleted after being read. Default is \code{FALSE}.}

\item{user}{Character. Your ENTSO-E API key. Passed to \code{entsoe_download_file} for authenticating the download request. Default is \code{Sys.getenv('ENTSOE_USER')}.}

\item{password}{Character. Your ENTSO-E API key. Passed to \code{entsoe_download_file} for authenticating the download request. Default is \code{Sys.getenv('ENTSOE_PASSWORD')}.}

\item{output_folder}{Character. The folder where the raw CSV file will be saved. Default is \code{tempdir()}.}
}
\value{
A \code{data.table} in long format with the following columns:
\describe{
\item{DATE}{The date of the data point.}
\item{TIME}{The time of the data point in \code{HH:MM} format.}
\item{HOUR}{The hour extracted from the \code{DATETIME} column as an integer.}
\item{RESOLUTION}{The resolution code (e.g., \code{quarter-hour}, \code{hourly}).}
\item{CODE_MAP}{The map code (e.g., country, bidding zone).}
\item{CODE_EIC}{The EIC code for the area (e.g., grid or market region).}
\item{PRODUCTION_TYPE}{The type of power generation (e.g., "Fossil Oil", "Wind Onshore").}
\item{ASSET_CATEGORY}{The category of the asset (e.g., "Thermal", "Wind", etc.).}
\item{ACTUAL_GENERATION}{The actual generation output for the given time period.}
\item{ACTUAL_CONSUMPTION}{The actual consumption for the given time period.}
\item{UPDATETIME}{The last update time for the data entry.}
\item{VARIABLE}{The variable being measured, either "ACTUAL_GENERATION" or "ACTUAL_CONSUMPTION".}
\item{VALUE}{The value for the corresponding variable.}
}
}
\description{
This function retrieves and processes the ENTSO-E actual generation data for a specified year and month.
It supports an option to either retain or delete the raw CSV file after processing based on the \code{raw} argument.
It also allows specifying the output folder for the downloaded CSV file.
}
\examples{
\dontrun{
# Download and process the data for October 2024, keeping the CSV file
result_raw <- entsoe_actual_generation(2024, 10, raw = TRUE)

# Download and process the data for October 2024, deleting the CSV file after processing
result <- entsoe_actual_generation(2024, 10)
}

}
