% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_prices_dam.R
\name{entsoe_dam_prices}
\alias{entsoe_dam_prices}
\title{Retrieve Day-Ahead Market Prices for a Specific Country}
\usage{
entsoe_dam_prices(
  country,
  from_data,
  to_data,
  api_key = Sys.getenv("ENTSOE_KEY")
)
}
\arguments{
\item{country}{A character string specifying the country for which to retrieve DAM prices. The country should match an entry in the \code{entsoe_countries} table.}

\item{from_data}{A Date object specifying the start date for the data range. Default is 365 days before the current date.}

\item{to_data}{A Date object specifying the end date for the data range. Default is the current date.}

\item{api_key}{A character string containing the ENTSOE API key. It can be passed directly or retrieved from the environment variable \code{ENTSOE_KEY}.}
}
\value{
A \code{data.table} containing the formatted DAM price data for the specified country and time range. The data table includes:
\itemize{
\item \code{DATE}: The date in \code{YYYY-MM-DD} format.
\item \code{HOUR}: The hour of the day (position in the data).
\item \code{VALUE}: The price value in EUR.
\item \code{UNIT}: The unit of the price (always 'EUR').
}
}
\description{
This function retrieves Day-Ahead Market (DAM) prices for a specific country from the ENTSOE Transparency Platform API.
The function fetches market price data for a given time range and country, processes the data, and formats it into a \code{data.table}.
}
\examples{
# Example usage to retrieve DAM prices for Italy (North)
country <- "Italy (North)"
from_data <- Sys.Date() - 365
to_data <- Sys.Date()
dam_prices <- entsoe_dam_prices(country, from_data, to_data, api_key = Sys.getenv('ENTSOE_KEY'))
print(dam_prices)

}
